requirements.txt
alembic==1.5.5
aniso8601==9.0.1
bcrypt==3.2.0
cffi==1.14.5
click==7.1.2
Flask==1.1.2
Flask-Bcrypt==0.7.1
Flask-Cors==3.0.10
Flask-DB==0.3.1
Flask-Migrate==2.7.0
Flask-RESTful==0.3.8
Flask-Script==2.0.6
Flask-SQLAlchemy==2.4.4
gunicorn==20.0.4
itsdangerous==1.1.0
Jinja2==2.11.3
ma==1.0.2
ma.mongo==1.0.2
Mako==1.1.4
MarkupSafe==1.1.1
marshmallow==3.10.0
psycopg2==2.8.6
pycparser==2.20
PyJWT==2.0.1
python-dateutil==2.8.1
python-dotenv==0.15.0
python-editor==1.0.4
pytz==2021.1
six==1.15.0
SQLAlchemy==1.3.23
SQLAlchemy-Utils==0.36.8
Werkzeug==1.0.1
